Update 作業練習 變數和資料類型 修改

+++ Assets/Scripts/Practice_2_DateType.cs
@@ -5,5 +5,5 @@ public class Practice_2_DateType :MonoBehaviour
 {
     public float i1 = 7.7f;
     public double i2 = 0.12345f;
-    public decimal i3 = (decimal)1.23456789f;
+    public decimal i3 = 1.23456789m;
 }
